% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{auto_determine_del_col}
\alias{auto_determine_del_col}
\title{Helper function to auto-determine columns to be deleted}
\usage{
auto_determine_del_col(d)
}
\arguments{
\item{data.frame}{from which columns are to be deleted}
}
\value{
vector with number of columns to be deleted
}
\description{
Helper function to auto-determine columns to be deleted
}
