% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing}
\alias{check_missing}
\title{A utility function reporting if the information the user requested
can indeed be found in the file.}
\usage{
check_missing(x, string, name)
}
\arguments{
\item{x}{the info requested by the user, e.g. the name of Ibex controller
or the element_number}

\item{string}{unique values in the corresponding column of the Ibex file
(column 3 for controller, column 5 for element_number )}

\item{name}{character, specifying the type of info we are checking for.
Will be inserted in the report message.}
}
\value{
invisibly, the number of requested pieces of info absent from the
        Ibex file.
}
\description{
A utility function reporting if the information the user requested
can indeed be found in the file.
}
