% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{count_columns}
\alias{count_columns}
\title{A function to count number of columns in a file.}
\usage{
count_columns(file_name)
}
\arguments{
\item{file_name}{character. Name fo the file where the columns should be counted}
}
\value{
numeric. Number of columns. If an error arises, assumes that there
        are 12 columns in the file.
}
\description{
A function to count number of columns in a file.
}
