% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{delete_columns}
\alias{delete_columns}
\title{A function to delete columns from the data based on data and/or user
requests.}
\usage{
delete_columns(d, del_col = NULL, del_mode = c("auto", "user", "mixed"),
  verbose = TRUE)
}
\arguments{
\item{d}{data.frame the columns of which are to be deleted}

\item{del_col}{numeric. Vector of column numbers to be deleted.}

\item{del_mode}{how to determine which columns have to be deleted. Can be
one of: \dQuote{auto} (decide based on data), \dQuote{user} (only take into account
user's requests), \dQuote{mixed} (take into account user request and data).
\dQuote{auto} is used by default if "del_col" is not specified; if it is, \dQuote{mixed}
is used. See Details}

\item{verbose}{logical. If \code{TRUE} (default), report which columns were
deleted and why.}
}
\value{
data.frame `d` with deleted columns
}
\description{
A function to delete columns from the data based on data and/or user
requests.
}
\details{
The first seven columns in IBEX results always are: \enumerate{ \item Time
the results are received \item MD5 hash of participant's IP \item Controller
type \item Item number \item Element number \item Type \item Group } By
default, the deletion mode is \dQuote{auto}. It means that the function will remove
columns not containing useful information. This means:

\itemize{ \item "md5 hash" column is always removed since it's not usually
used in analyses \item "controller" column is removed if there is just one
controller in the data - it is assumed, the user will know what controller
that is \item "element number" column is removed if that column only
contains zeros \item "group" column is removed if it only contains `NULL`
values }

The above applies to all controllers. In addition:

\itemize{ \item for DashedSentence and DashedAccpetabilityJudgment column 11
("line break") is removed if it only contains \code{FALSE} values. \item For
DashedAcceptabilityJudgment only column 15 ("is.correct") is removed if it
contains only \code{NULL} valuess (if this is the case, it means that correct
answers were not specified in IBEX, so this column is of no use). }

In \dQuote{user} deletion mode the data will not be taken into account at all, and
only the columns specified by user will be deleted.

In \dQuote{mixed} deletion mode the columns to delete are the union of the sets of
automatically deleted columns and user-specified columns.
}
