% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_sa_lines}
\alias{detect_sa_lines}
\title{Find which lines come from \dQuote{speeded acceptability} mode}
\usage{
detect_sa_lines(d)
}
\arguments{
\item{d}{\code{data.frame} with Ibex data which only contains rows coming
from \dQuote{DashedSentence} or \dQuote{DashedAcceptabilityJudgment}
controllers}
}
\value{
logical vector corresponding to rows in speeded acceptability mode.
}
\description{
It does so by checking whether column 11 is empty. Speeded acceptability
data only has 8 columns, so it will be true for it (and false for self-paced
reading data or question data).
}
