% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_sprt_lines}
\alias{detect_sprt_lines}
\title{Find which lines come from \dQuote{self-paced reading} mode}
\usage{
detect_sprt_lines(d)
}
\arguments{
\item{d}{\code{data.frame} with Ibex data which only contains rows coming
from \dQuote{DashedSentence} or \dQuote{DashedAcceptabilityJudgment}
controllers}
}
\value{
logical vector corresponding to rows in sprt mode.
}
\description{
It does so by taking column 10 and trying to convert it to numeric. For
data in \dQuote{self-paced reading} this column will contain RTs, so
the conversion will be successful.
}
