% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ibex}
\alias{format_ibex}
\title{Set the names and classes for the columns in Ibex data}
\usage{
format_ibex(d, col_names = NULL, partial_names = TRUE, col_classes = NULL,
  partial_classes = TRUE)
}
\arguments{
\item{d}{data.frame as returned by `format_ibex`}

\item{col_names}{character vector with names for the columns}

\item{partial_names}{logical. If `TRUE`, the first seven columns will receive
default names (`timestamp`, `md5_hash`, `controller`,
`item_number`, `element_number`, `type`, `group`),
since these columns are the same in all (default)
Ibex controllers. So `col_names` will be taken to
specify names for columns starting from 8.
If `FALSE`, `col_names` will be taken as specifying names
for all columns}

\item{col_classes}{character vector with classes to be assumed for the columns}

\item{partial_classes}{logical. If `TRUE`, the first seven columns will receive
                     default classes:

                     + timestamp, md5_hash, controller, type, group - `character`,
                     + item_number, element_number - `numeric`,

                     So `col_classes` will be taken to specify classes for columns
                     starting from 8. If `FALSE`, `col_classes` will be taken
                     as specifying classes for all columns}
}
\description{
Set the names and classes for the columns in Ibex data
}
