% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subj_info.R
\name{get_subj_info}
\alias{get_subj_info}
\title{Extract subjects info from Ibex data}
\usage{
get_subj_info(file_name, form_name = "intro", info = NULL,
  browser_info = FALSE, short_subj_ids = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{file_name}{character. Path to the file with Ibex results}

\item{form_name}{character. Useful if you have several Forms collecting
different info throughout the experiment. In this case form.name should
specify which of those Forms collected subject info you're interested in.
By default (in IBEX and here) it's \dQuote{intro}. You can also pass a vector
with several form names.}

\item{info}{character vector, describing which info about subjects you want
to extract (i.e. the names of the fields you specified in IBEX, e.g.
"age", "sex" etc). If \code{NULL} (deault), all the available info will be
extracted.}

\item{browser_info}{logical. Whether to extract subject's browser info or not
(defaults to \code{FALSE}).}

\item{short_subj_ids}{logical. If \code{FALSE}, just merge timestamp and md5 hash
together, If \code{TRUE} (default), add a new column with a simple index to
idenitfy subjects. One index is assigned to each combination of timestamp
and browser md5 hash. The column is put at the place of \dQuote{subj_uid}
column, and \dQuote{subj_uid} column is moved at the very end of the data.frame.}

\item{verbose}{logical. If \code{TRUE}, print the number of subjects and the kinds
of information extracted.}
}
\value{
data.frame with subject info
}
\description{
Extract subjects info from Ibex data
}
