% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{make_column_classes}
\alias{make_column_classes}
\title{A function to provide default names for columns and ensure that
every column has a name}
\usage{
make_column_classes(n_cols, col_classes = NULL, partial_classes = TRUE)
}
\arguments{
\item{n_cols}{\code{numeric}. Number of columns in the data}

\item{col_classes}{\code{character} vector with classes to be assumed for the columns.
Currently, the following classes are allowed: \dQuote{character}, \dQuote{numeric},
\dQuote{integer}, \dQuote{logical}, \dQuote{factor}. Additionally, one may
use \dQuote{asis}, if the class of a column should not be changed. If the class
for a column is not specified explicitly, \dQuote{asis} will be assumed.}

\item{partial_classes}{\code{logical}. If \code{TRUE}, the first seven columns will receive
                     default classes:

                     + subject - \code{factor}
                     + md5_hash, controller, type, item - \code{character}
                      (notice that in Ibex items can be identified with character
                      strings, that's why item is not \code{numeric} by default),
                     + presentation_order, element_number - \code{numeric},

                     So \code{col_names} will be taken to specify classes for columns
                     starting from 8. If \code{FALSE}, \code{col_classes} will be taken
                     as specifying classes for all columns. If there are less
                     than 8 columns, \code{partial classes} will always be taken
                     to be \code{FALSE}.}
}
\value{
vector with column classes
}
\description{
A function to provide default names for columns and ensure that
every column has a name
}
