% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{make_column_names}
\alias{make_column_names}
\title{A function to provide default names for columns and ensure that
every column has a name}
\usage{
make_column_names(n_cols, col_names = NULL, partial_names = TRUE)
}
\arguments{
\item{n_cols}{numeric. Number of columns in the data}

\item{col_names}{character vector with names for the columns.}

\item{partial_names}{logical. If \code{TRUE}, the first seven columns will receive
default names (\dQuote{subject}, \dQuote{md5_hash}, \dQuote{controller},
\dQuote{presentation_order}, \dQuote{element_number},
\dQuote{type}, \dQuote{item}),
since these columns are the same in all (default)
Ibex controllers. So \code{col_names} will be taken to
specify names for columns starting from 8.
If \code{FALSE}, \code{col_names} will be taken as specifying names
for all columns. If there are fewer than 8 columns,
\code{partial_names} will always be taken to be \code{FALSE}.}
}
\value{
vector with column names
}
\description{
A function to provide default names for columns and ensure that
every column has a name
}
