% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ibex}
\alias{read_ibex}
\title{Read the csv file with Ibex results}
\usage{
read_ibex(file_name, ...)
}
\arguments{
\item{file_name}{character. Path to the file with Ibex results}
}
\value{
data frame with the contents of the file
}
\description{
Thin wrapper around `read.csv`
}
