% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_subjects}
\alias{recode_subjects}
\title{Set unique subject IDs}
\usage{
recode_subjects(d, short_ids = TRUE)
}
\arguments{
\item{d}{data.frame with (a subset of) Ibex data}

\item{short_ids}{logical. If `FALSE`, just merge timestamp and md5 hash
together, If `TRUE` (default), add a new column with a simple index to
idenitfy subjects. One index is assigned to each combination of timestamp
and browser md5 hash. The column is put at the place of `subj_uid`
column, and `subj_uid` column is moved at the very end of the data.frame
(in order not to interfere with column deletion; otherwise we would have
to offset the indices of columns specified for deletion by 1).}
}
\value{
data.frame with updated subject codes
}
\description{
Ibex provides 2 pieces of information identifying a participant: timestamp
of the moment when the results and md5 hash of browser and system
information. Neither of these parameters in itself is guaranteed to be
unique; their combination has more chances of being so.
}
\details{
This function merges the timestamp and md5 into a single string (separated
by "_"), and puts it into `subj_uid` column (which replaces the `timestamp`
column. If necessary, timestamp can be extracted by simply breaking the
combined string at "_").
}
