% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{report_del_col}
\alias{report_del_col}
\title{Helper function to report which columns were deleted}
\usage{
report_del_col(del_col, d_colnames)
}
\arguments{
\item{del_col}{Vector of column numbers to be deleted.}

\item{d_colnames}{Vector of columns names (used to report more informatively
which columns were deleted)}
}
\value{
data.frame with deleted columns info
}
\description{
Helper function to report which columns were deleted
}
