% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{set_column_classes}
\alias{set_column_classes}
\title{A function to change classes of multiple data.frame columns}
\usage{
set_column_classes(d, col_classes)
}
\arguments{
\item{d}{data.frame}

\item{col_classes}{character vector with classes to be assumed for the columns}
}
\value{
data.frame with updated column classes
}
\description{
A function to change classes of multiple data.frame columns
}
