% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_ibex}
\alias{subset_ibex}
\title{Select the right subset of the Ibex data}
\usage{
subset_ibex(d, controller, elem_number = NULL)
}
\arguments{
\item{d}{data frame with Ibex results as returned by `read_ibex`}

\item{controller}{character. Name of Ibex controller to get the data for.
In principle can be a vector, but it is recommended to use single values,
since often different controllers have a different number of columns, and
the content of the columns beyond the first seven is also
controller-specific.}

\item{elem_number}{Element number in IBEX. Only useful when a) You
specified multiple controllers within one item and b) Some of those
controllers have the same type. In this case you can specify which
exactly of those duplicate controllers you need.}
}
\value{
data frame with data for requested controllers / elements
}
\description{
Select the right subset of the Ibex data
}
